requirements.txt
requests>=2.31.0
openpyxl>=3.1.0
Pillow>=10.0.0
aiohttp>=3.8.0
supabase>=2.3.5
python-dotenv>=1.0.0
pandas>=2.0.0
